cleaned up code and changed button
the show purple images button had a click event and now it has a hover event. i changed it because i didn't want users to be able to move the purple images like they were the images of regular pieces. now if the user tries to move the mouse off the button to move a purple image the images will disappear.